auth, route protection

+++ src/controllers/cardsController.ts
@@ -1,8 +1,9 @@
 import { NextFunction, Request, Response } from 'express';
-import RequestWithUserRole from '../models/request';
+import RequestWithUser from '../models/request';
 import Card from '../models/card';
 import handleError from '../middlwares/handleError';
 import NotFoundError from '../errors/notFoundError';
+import { ForbiddenError, NotAuthorizedError } from '../errors';
 
 async function getCardById(_id: string | undefined) {
   const user = await Card.findById(_id);
@@ -21,20 +22,29 @@ export const getCards = async (req: Request, res: Response, next: NextFunction) 
   }
 };
 
-export const createCard = async (req: RequestWithUserRole, res: Response, next: NextFunction) => {
+export const createCard = async (req: RequestWithUser, res: Response, next: NextFunction) => {
   try {
+    if (!req.user || !req.user._id) {
+      throw new NotAuthorizedError();
+    }
     const { name, link } = req.body;
-    const card = await (await Card.create({ name, link, owner: req.user })).populate('owner');
+    const card = await (await Card.create({ name, link, owner: req.user._id })).populate('owner');
     res.status(200).send({ data: card });
   } catch (error: any) {
     handleError(error, req, res, next);
   }
 };
 
-export const deleteCard = async (req: Request, res: Response, next: NextFunction) => {
+export const deleteCard = async (req: RequestWithUser, res: Response, next: NextFunction) => {
   try {
+    console.log(req.user);
+
+    if (!req.user || !req.user._id) {
+      throw new NotAuthorizedError();
+    }
     const { cardId } = req.params;
     const card = await getCardById(cardId);
+    if (card.owner.toString() !== req.user._id) throw new ForbiddenError();
     await card.delete();
     res.status(200).send({ message: 'Пост удален' });
   } catch (error) {
@@ -42,7 +52,7 @@ export const deleteCard = async (req: Request, res: Response, next: NextFunction
   }
 };
 
-export const putLike = async (req: RequestWithUserRole, res: Response, next: NextFunction) => {
+export const putLike = async (req: RequestWithUser, res: Response, next: NextFunction) => {
   try {
     const { cardId } = req.params;
     const card = await Card.findByIdAndUpdate(
@@ -59,7 +69,7 @@ export const putLike = async (req: RequestWithUserRole, res: Response, next: Nex
   }
 };
 
-export const deleteLike = async (req: RequestWithUserRole, res: Response, next: NextFunction) => {
+export const deleteLike = async (req: RequestWithUser, res: Response, next: NextFunction) => {
   try {
     const { cardId } = req.params;
     const card = await Card.findByIdAndUpdate(
